Finished POC adaptor
Currently relies on user clicking somewhere to pop up dialogue.

Note that there is no way to dynamically enable/disable this adaptor.

Also, loading a directory is not currently supported. In the future, we
will need a minimal wiki "reader" which only contains the skinny
tiddlers and this plugin. We might be able to avoid the skinny tiddlers
by adding a login method?