Enable integrity check
Also have an option for saving file handles even if served from
file://. This is a big time saver and I think technically savvy users
should be allowed to turn this off.

This integrity check is done by comparing the hash of the file on disk
with the previous hash which was saved. We use a modal to force people
to choose a location quickly to minimize the chance of conflicts.

We wait until after the initial save to do the integrity check in order
to make saving over an existing file easier if that is desired. This
behavior might change in the future.